changed database structure for a much simpler one removing intermediate links table as data is only designed to be accessed in the direction user -> sessions adapted repo to change repo test-coverage completed

+++ app/src/main/java/fr/shining_cat/simplehiit/domain/datainterfaces/SimpleHiitRepository.kt
@@ -13,6 +13,6 @@ interface SimpleHiitRepository {
     suspend fun updateUser(user:User):Output<Int>
     suspend fun deleteUser(user:User):Output<Int>
     //
-    suspend fun insertSession(session: Session): Output<Long>
+    suspend fun insertSession(session: Session): Output<Int>
     suspend fun getSessionsForUser(user: User):Output<List<Session>>
 }
